categories Crud done

+++ DevDiary/Data/DataSeeding.cs
@@ -8,16 +8,22 @@ public static class DataSeeding
 
     public static async Task Initialize(ApplicationDbContext context)
     {
-        List<string> Categories = ["C# Foundation",
+        var diaryCategory = context?.DiaryCategories?.FirstOrDefault();
+        if (diaryCategory is null)
+        {
+
+            List<string> Categories = ["C# Foundation",
              "Entity Framework",
              "JavaScript","React Foundation","SQL Foundation", "React Hook", "Desgin Patterns"];
-        List<DiaryCategory> diaryCategories = [];
-        foreach (var category in Categories)
-        {
-            diaryCategories.Add(new DiaryCategory { Name = category });
+            List<DiaryCategory> diaryCategories = [];
+            foreach (var category in Categories)
+            {
+                diaryCategories.Add(new DiaryCategory { Name = category });
+            }
+            Console.WriteLine(String.Join(',', diaryCategories.Select(x => x.Id)));
+            await context.Database.ExecuteSqlRawAsync("delete from diaryCategories");
+            await context.DiaryCategories.AddRangeAsync(diaryCategories);
+            await context.SaveChangesAsync();
         }
-        await context.Database.ExecuteSqlRawAsync("delete from diaryCategories");
-        await context.DiaryCategories.AddRangeAsync(diaryCategories);
-        await context.SaveChangesAsync();
     }
 }
